refactor(rope): split rope force to physics
